chore(mobile-header): add scenario for floating header
