+++ src/mobile-header/__tests__/mobile-header.stories.tsx
@@ -5,6 +5,8 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 */
 import React from 'react';
-import { Scenario as MobileHeaderDefault } from './mobile-header.scenario';
+import { Scenario as MobileHeaderFixed } from './mobile-header-fixed.scenario';
+import { Scenario as MobileHeaderFloating } from './mobile-header-floating.scenario';
 
-export const MobileHeader = () => <MobileHeaderDefault />;
+export const Fixed = () => <MobileHeaderFixed />;
+export const Floating = () => <MobileHeaderFloating />;
